SurveyResult 를 별도 파일로 옮김
